requirements.txt
openai
chromadb
python-dotenv
Flask
requests
PyPDF2
python-docx
gunicorn